refactor: User에 있는 정보 owner와 student로 리팩토링
