+++ src/main/java/in/koreatech/koin/admin/user/controller/AdminUserController.java
@@ -52,28 +52,6 @@ public class AdminUserController implements AdminUserApi{
 
     private final AdminUserService adminUserService;
 
-    @PutMapping("/admin/owner/{id}/authed")
-    public ResponseEntity<Void> allowOwnerPermission(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId) {
-        adminUserService.allowOwnerPermission(id);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/admin/students")
-    public ResponseEntity<AdminStudentsResponse> getStudents(
-        @RequestParam(required = false) Integer page,
-        @RequestParam(required = false) Integer limit,
-        @RequestParam(required = false) Boolean isAuthed,
-        @RequestParam(required = false) String nickname,
-        @RequestParam(required = false) String email,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        StudentsCondition studentsCondition = new StudentsCondition(page, limit, isAuthed, nickname, email);
-        AdminStudentsResponse adminStudentsResponse = adminUserService.getStudents(studentsCondition);
-        return ResponseEntity.ok().body(adminStudentsResponse);
-    }
-
     @PostMapping("/admin")
     public ResponseEntity<Void> createAdmin(
         @RequestBody @Valid CreateAdminRequest request,
@@ -168,60 +146,6 @@ public class AdminUserController implements AdminUserApi{
     ) {
         adminUserService.updateAdminPermission(request, id, adminId);
         return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/admin/users/student/{id}")
-    public ResponseEntity<AdminStudentResponse> getStudent(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminStudentResponse adminStudentResponse = adminUserService.getStudent(id);
-        return ResponseEntity.ok().body(adminStudentResponse);
-    }
-
-    @PutMapping("/admin/users/student/{id}")
-    public ResponseEntity<AdminStudentUpdateResponse> updateStudent(
-        @Valid @RequestBody AdminStudentUpdateRequest adminRequest,
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminStudentUpdateResponse adminStudentUpdateResponse = adminUserService.updateStudent(id, adminRequest);
-        return ResponseEntity.ok().body(adminStudentUpdateResponse);
-    }
-
-    @GetMapping("/admin/users/owner/{id}")
-    public ResponseEntity<AdminOwnerResponse> getOwner(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminOwnerResponse adminOwnerResponse = adminUserService.getOwner(id);
-        return ResponseEntity.ok().body(adminOwnerResponse);
-    }
-
-    @PutMapping("/admin/users/owner/{id}")
-    public ResponseEntity<AdminOwnerUpdateResponse> updateOwner(
-        @PathVariable Integer id,
-        @RequestBody @Valid AdminOwnerUpdateRequest request,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminOwnerUpdateResponse adminOwnerUpdateResponse = adminUserService.updateOwner(id, request);
-        return ResponseEntity.ok().body(adminOwnerUpdateResponse);
-    }
-
-    @GetMapping("/admin/users/new-owners")
-    public ResponseEntity<AdminNewOwnersResponse> getNewOwners(
-        @ParameterObject @ModelAttribute OwnersCondition ownersCondition,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        return ResponseEntity.ok().body(adminUserService.getNewOwners(ownersCondition));
-    }
-
-    @GetMapping("/admin/users/owners")
-    public ResponseEntity<AdminOwnersResponse> getOwners(
-        @ParameterObject @ModelAttribute OwnersCondition ownersCondition,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        return ResponseEntity.ok().body(adminUserService.getOwners(ownersCondition));
     }
 
     @GetMapping("/admin/users/{id}")
