+++ src/main/java/in/koreatech/koin/admin/user/service/AdminUserService.java
@@ -3,60 +3,34 @@ package in.koreatech.koin.admin.user.service;
 import static in.koreatech.koin.domain.user.model.UserType.ADMIN;
 
 import java.time.LocalDateTime;
-import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import in.koreatech.koin.admin.shop.repository.AdminShopRepository;
 import in.koreatech.koin.admin.user.dto.AdminLoginRequest;
 import in.koreatech.koin.admin.user.dto.AdminLoginResponse;
-import in.koreatech.koin.admin.user.dto.AdminNewOwnersResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnerResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnerUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminOwnerUpdateResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnersResponse;
 import in.koreatech.koin.admin.user.dto.AdminPasswordChangeRequest;
 import in.koreatech.koin.admin.user.dto.AdminPermissionUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminStudentUpdateResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentsResponse;
 import in.koreatech.koin.admin.user.dto.AdminTokenRefreshRequest;
 import in.koreatech.koin.admin.user.dto.AdminTokenRefreshResponse;
 import in.koreatech.koin.admin.user.dto.AdminUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminsCondition;
 import in.koreatech.koin.admin.user.dto.AdminsResponse;
 import in.koreatech.koin.admin.user.dto.CreateAdminRequest;
-import in.koreatech.koin.admin.user.dto.OwnersCondition;
-import in.koreatech.koin.admin.user.dto.StudentsCondition;
 import in.koreatech.koin.admin.user.model.Admin;
-import in.koreatech.koin.admin.user.repository.AdminOwnerRepository;
-import in.koreatech.koin.admin.user.repository.AdminOwnerShopRedisRepository;
+import in.koreatech.koin.admin.owner.repository.AdminOwnerRepository;
 import in.koreatech.koin.admin.user.repository.AdminRepository;
-import in.koreatech.koin.admin.user.repository.AdminStudentRepository;
+import in.koreatech.koin.admin.student.repository.AdminStudentRepository;
 import in.koreatech.koin.admin.user.repository.AdminTokenRepository;
 import in.koreatech.koin.admin.user.repository.AdminUserRepository;
-import in.koreatech.koin.domain.owner.model.Owner;
-import in.koreatech.koin.domain.owner.model.OwnerIncludingShop;
-import in.koreatech.koin.domain.owner.model.OwnerShop;
-import in.koreatech.koin.domain.shop.model.shop.Shop;
-import in.koreatech.koin.domain.student.exception.StudentDepartmentNotValidException;
-import in.koreatech.koin.domain.student.model.Student;
-import in.koreatech.koin.domain.student.model.StudentDepartment;
-import in.koreatech.koin.domain.user.exception.DuplicationNicknameException;
 import in.koreatech.koin.domain.user.exception.UserNotFoundException;
 import in.koreatech.koin.domain.user.model.User;
-import in.koreatech.koin.domain.user.model.UserGender;
 import in.koreatech.koin.domain.user.model.UserToken;
 import in.koreatech.koin.domain.user.model.UserType;
 import in.koreatech.koin.global.auth.JwtProvider;
@@ -75,22 +49,11 @@ public class AdminUserService {
     private final JwtProvider jwtProvider;
     private final AdminStudentRepository adminStudentRepository;
     private final AdminOwnerRepository adminOwnerRepository;
-    private final AdminOwnerShopRedisRepository adminOwnerShopRedisRepository;
     private final AdminUserRepository adminUserRepository;
-    private final AdminShopRepository adminShopRepository;
     private final PasswordEncoder passwordEncoder;
     private final AdminTokenRepository adminTokenRepository;
     private final AdminRepository adminRepository;
 
-    public AdminStudentsResponse getStudents(StudentsCondition studentsCondition) {
-        Integer totalStudents = adminStudentRepository.findAllStudentCount();
-        Criteria criteria = Criteria.of(studentsCondition.page(), studentsCondition.limit(), totalStudents);
-
-        PageRequest pageRequest = PageRequest.of(criteria.getPage(), criteria.getLimit());
-        Page<Student> studentsPage = adminStudentRepository.findByConditions(studentsCondition, pageRequest);
-
-        return AdminStudentsResponse.from(studentsPage);
-    }
 
     @Transactional
     public AdminResponse createAdmin(CreateAdminRequest request, Integer adminId) {
@@ -230,145 +193,6 @@ public class AdminUserService {
         }
 
         adminRepository.getById(id).updatePermission(request);
-    }
-
-    @Transactional
-    public void allowOwnerPermission(Integer id) {
-        Owner owner = adminOwnerRepository.getById(id);
-        owner.getUser().auth();
-        Optional<OwnerShop> ownerShop = adminOwnerShopRedisRepository.findById(id);
-        if (ownerShop.isPresent()) {
-            Integer shopId = ownerShop.get().getShopId();
-            if (shopId != null) {
-                Shop shop = adminShopRepository.getById(shopId);
-                shop.updateOwner(owner);
-                owner.setGrantShop(true);
-            }
-            adminOwnerShopRedisRepository.deleteById(id);
-        }
-    }
-
-    public AdminStudentResponse getStudent(Integer userId) {
-        Student student = adminStudentRepository.getById(userId);
-        return AdminStudentResponse.from(student);
-    }
-
-    @Transactional
-    public AdminStudentUpdateResponse updateStudent(Integer id, AdminStudentUpdateRequest adminRequest) {
-        Student student = adminStudentRepository.getById(id);
-        User user = student.getUser();
-        validateNicknameDuplication(adminRequest.nickname(), id);
-        validateDepartmentValid(adminRequest.major());
-        user.update(adminRequest.nickname(), adminRequest.name(),
-            adminRequest.phoneNumber(), UserGender.from(adminRequest.gender()));
-        user.updateStudentPassword(passwordEncoder, adminRequest.password());
-        student.update(adminRequest.studentNumber(), adminRequest.major());
-        adminStudentRepository.save(student);
-
-        return AdminStudentUpdateResponse.from(student);
-    }
-
-    public AdminNewOwnersResponse getNewOwners(OwnersCondition ownersCondition) {
-        ownersCondition.checkDataConstraintViolation();
-
-        Integer totalOwners = adminUserRepository.findUsersCountByUserTypeAndIsAuthed(UserType.OWNER, false);
-        Criteria criteria = Criteria.of(ownersCondition.page(), ownersCondition.limit(), totalOwners);
-        Sort.Direction direction = ownersCondition.getDirection();
-
-        Page<Owner> result = getNewOwnersResultPage(ownersCondition, criteria, direction);
-
-        List<OwnerIncludingShop> ownerIncludingShops = result.getContent().stream()
-            .map(owner -> {
-                Optional<OwnerShop> ownerShop = adminOwnerShopRedisRepository.findById(owner.getId());
-                return ownerShop
-                    .map(os -> {
-                        Shop shop = adminShopRepository.findById(os.getShopId()).orElse(null);
-                        return OwnerIncludingShop.of(owner, shop);
-                    })
-                    .orElseGet(() -> OwnerIncludingShop.of(owner, null));
-            })
-            .collect(Collectors.toList());
-
-        return AdminNewOwnersResponse.of(ownerIncludingShops, result, criteria);
-    }
-
-    public AdminOwnersResponse getOwners(OwnersCondition ownersCondition) {
-        ownersCondition.checkDataConstraintViolation();
-
-        Integer totalOwners = adminUserRepository.findUsersCountByUserTypeAndIsAuthed(UserType.OWNER, true);
-        Criteria criteria = Criteria.of(ownersCondition.page(), ownersCondition.limit(), totalOwners);
-        Sort.Direction direction = ownersCondition.getDirection();
-
-        Page<Owner> result = getOwnersResultPage(ownersCondition, criteria, direction);
-
-        return AdminOwnersResponse.of(result, criteria);
-    }
-
-    private Page<Owner> getOwnersResultPage(OwnersCondition ownersCondition, Criteria criteria,
-        Sort.Direction direction) {
-        PageRequest pageRequest = PageRequest.of(criteria.getPage(), criteria.getLimit(),
-            Sort.by(direction, "user.createdAt"));
-
-        Page<Owner> result;
-
-        if (ownersCondition.searchType() == OwnersCondition.SearchType.EMAIL) {
-            result = adminOwnerRepository.findPageOwnersByEmail(ownersCondition.query(), pageRequest);
-        } else if (ownersCondition.searchType() == OwnersCondition.SearchType.NAME) {
-            result = adminOwnerRepository.findPageOwnersByName(ownersCondition.query(), pageRequest);
-        } else {
-            result = adminOwnerRepository.findPageOwners(pageRequest);
-        }
-
-        return result;
-    }
-
-    private Page<Owner> getNewOwnersResultPage(OwnersCondition ownersCondition, Criteria criteria,
-        Sort.Direction direction) {
-        PageRequest pageRequest = PageRequest.of(criteria.getPage(), criteria.getLimit(),
-            Sort.by(direction, "user.createdAt"));
-
-        Page<Owner> result;
-
-        if (ownersCondition.searchType() == OwnersCondition.SearchType.EMAIL) {
-            result = adminOwnerRepository.findPageUnauthenticatedOwnersByEmail(ownersCondition.query(), pageRequest);
-        } else if (ownersCondition.searchType() == OwnersCondition.SearchType.NAME) {
-            result = adminOwnerRepository.findPageUnauthenticatedOwnersByName(ownersCondition.query(), pageRequest);
-        } else {
-            result = adminOwnerRepository.findPageUnauthenticatedOwners(pageRequest);
-        }
-
-        return result;
-    }
-
-    private void validateNicknameDuplication(String nickname, Integer userId) {
-        if (nickname != null &&
-            adminUserRepository.existsByNicknameAndIdNot(nickname, userId)) {
-            throw DuplicationNicknameException.withDetail("nickname : " + nickname);
-        }
-    }
-
-    private void validateDepartmentValid(String department) {
-        if (department != null && !StudentDepartment.isValid(department)) {
-            throw StudentDepartmentNotValidException.withDetail("학부(학과) : " + department);
-        }
-    }
-
-    public AdminOwnerResponse getOwner(Integer ownerId) {
-        Owner owner = adminOwnerRepository.getById(ownerId);
-
-        List<Integer> shopsId = adminShopRepository.findAllByOwnerId(ownerId)
-            .stream()
-            .map(Shop::getId)
-            .toList();
-
-        return AdminOwnerResponse.of(owner, shopsId);
-    }
-
-    @Transactional
-    public AdminOwnerUpdateResponse updateOwner(Integer ownerId, AdminOwnerUpdateRequest request) {
-        Owner owner = adminOwnerRepository.getById(ownerId);
-        owner.update(request);
-        return AdminOwnerUpdateResponse.from(owner);
     }
 
     @Transactional
